The existing mpi parameters provided in v2.0 do not specify any binding parameters. As a result, depending on the configuration of your environment and test you may expect one of three behaviors from the openmpi defaults:
- Bind to core: when the number of processes is <= 2
- Bind to package: when the number of processes is > 2
- Bind to none: when oversubscribed

While the existing benchmark does offer the oversubscribe flag to disable binding, it does not resolve the map-by issue and is not particularly helpful in smaller configs.

Additionally, the default map-by setting is by "slot" which can lead to assymmetric distribution of processes if not careful.

This commit modifies the mpi parameters to a defined default that varies with node count:

if nodes = 1, we bind-to none map-by socket
If nodes >1, we bind-to none map-by node

On a single client configuration, this helps by giving flexibility for OS to schedule processes and prioritizes socket locality.
On multi-node configs, this ensures a more symmetric distribution of processes across nodes.

In both scenarios, this should minimize the chance of client cpu bottlenecks artificieally limiting performance.